III-4475 - Extract organizer mutations & logic to seperate OrganizerStep component
